Issue #86 fix: 레포지토리 분리
