+++ app/src/main/java/com/lateinit/rightweight/data/repository/RoutineRepository.kt
@@ -6,9 +6,6 @@ import com.lateinit.rightweight.data.database.entity.ExerciseSet
 import com.lateinit.rightweight.data.database.entity.Routine
 import com.lateinit.rightweight.data.database.intermediate.DayWithExercises
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
-import com.lateinit.rightweight.data.remote.model.DayField
-import com.lateinit.rightweight.data.remote.model.ExerciseField
-import com.lateinit.rightweight.data.remote.model.ExerciseSetField
 
 interface RoutineRepository {
 
@@ -31,20 +28,5 @@ interface RoutineRepository {
     suspend fun getRoutineWithDaysByRoutineId(routineId: String): RoutineWithDays
     suspend fun getDayWithExercisesByDayId(dayId: String): DayWithExercises
     suspend fun removeRoutineById(routineId: String)
-    suspend fun shareRoutine(userId: String, routineId: String, routine: Routine)
-    suspend fun shareDay(routineId: String, dayId: String, dayField: DayField)
-    suspend fun shareExercise(
-        routineId: String,
-        dayId: String,
-        exerciseId: String,
-        exerciseField: ExerciseField
-    )
 
-    suspend fun shareExerciseSet(
-        routineId: String,
-        dayId: String,
-        exerciseId: String,
-        exerciseSetId: String,
-        exerciseSetField: ExerciseSetField
-    )
 }
