+++ app/src/main/java/com/lateinit/rightweight/data/repository/RoutineRepositoryImpl.kt
@@ -7,17 +7,10 @@ import com.lateinit.rightweight.data.database.entity.Routine
 import com.lateinit.rightweight.data.database.intermediate.DayWithExercises
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
 import com.lateinit.rightweight.data.datasource.RoutineLocalDataSource
-import com.lateinit.rightweight.data.datasource.RoutineRemoteDataSource
-import com.lateinit.rightweight.data.remote.model.DayField
-import com.lateinit.rightweight.data.remote.model.ExerciseField
-import com.lateinit.rightweight.data.remote.model.ExerciseSetField
-import com.lateinit.rightweight.data.remote.model.RootField
-import com.lateinit.rightweight.util.toSharedRoutineField
 import javax.inject.Inject
 
 class RoutineRepositoryImpl @Inject constructor(
-    private val routineLocalDataSource: RoutineLocalDataSource,
-    private val routineRemoteDataSource: RoutineRemoteDataSource
+    private val routineLocalDataSource: RoutineLocalDataSource
 ) : RoutineRepository {
 
     override suspend fun insertRoutine(
@@ -75,39 +68,5 @@ class RoutineRepositoryImpl @Inject constructor(
         routineLocalDataSource.removeRoutineById(routineId)
     }
 
-    override suspend fun shareRoutine(userId: String, routineId: String, routine: Routine) {
-        routineRemoteDataSource.shareRoutine(
-            routineId, RootField(routine.toSharedRoutineField(userId))
-        )
-    }
-
-    override suspend fun shareDay(routineId: String, dayId: String, dayField: DayField) {
-        routineRemoteDataSource.shareDay(
-            routineId, dayId, RootField(dayField)
-        )
-    }
-
-    override suspend fun shareExercise(
-        routineId: String,
-        dayId: String,
-        exerciseId: String,
-        exerciseField: ExerciseField
-    ) {
-        routineRemoteDataSource.shareExercise(
-            routineId, dayId, exerciseId, RootField(exerciseField)
-        )
-    }
-
-    override suspend fun shareExerciseSet(
-        routineId: String,
-        dayId: String,
-        exerciseId: String,
-        exerciseSetId: String,
-        exerciseSetField: ExerciseSetField
-    ) {
-        routineRemoteDataSource.shareExerciseSet(
-            routineId, dayId, exerciseId, exerciseSetId, RootField(exerciseSetField)
-        )
-    }
 
 }
